--- a/.buildkite/pipeline.yml
+++ b/.buildkite/pipeline.yml
@@ -52,13 +52,13 @@ steps:
           - "**/build/reports/detekt/detekt.html"
 
       - label: "🕵️ Lint WordPress"
-        command: ".buildkite/commands/lint.sh wordpress"
+        command: ".buildkite/commands/lint.sh" "wordpress"
         plugins: [$CI_TOOLKIT]
         artifact_paths:
           - "**/build/reports/lint-results*.*"
 
       - label: "🕵️ Lint Jetpack"
-        command: ".buildkite/commands/lint.sh jetpack"
+        command: ".buildkite/commands/lint.sh" "jetpack"
         plugins: [$CI_TOOLKIT]
         artifact_paths:
           - "**/build/reports/lint-results*.*"
@@ -75,7 +75,7 @@ steps:
   - group: "🔬 Unit Tests"
     steps:
       - label: "🔬 Unit Test WordPress"
-        command: ".buildkite/commands/run-unit-tests.sh wordpress"
+        command: ".buildkite/commands/run-unit-tests.sh" "wordpress"
         plugins:
           - $CI_TOOLKIT
           - $TEST_COLLECTOR :
@@ -85,7 +85,7 @@ steps:
           - "**/build/test-results/merged-test-results.xml"
 
       - label: "🔬 Unit Test Processors"
-        command: ".buildkite/commands/run-unit-tests.sh processors"
+        command: ".buildkite/commands/run-unit-tests.sh" "processors"
         plugins:
           - $CI_TOOLKIT
           - $TEST_COLLECTOR :
@@ -95,7 +95,7 @@ steps:
           - "**/build/test-results/merged-test-results.xml"
 
       - label: "🔬 Unit Test Image Editor"
-        command: ".buildkite/commands/run-unit-tests.sh image-editor"
+        command: ".buildkite/commands/run-unit-tests.sh" "image-editor"
         plugins:
           - $CI_TOOLKIT
           - $TEST_COLLECTOR :
@@ -105,7 +105,7 @@ steps:
           - "**/build/test-results/merged-test-results.xml"
 
       - label: "🔬 Unit Test FluxC"
-        command: ".buildkite/commands/run-unit-tests.sh fluxc"
+        command: ".buildkite/commands/run-unit-tests.sh" "fluxc"
         plugins:
           - $CI_TOOLKIT
           - $TEST_COLLECTOR:
@@ -115,7 +115,7 @@ steps:
           - "**/build/test-results/merged-test-results.xml"
 
       - label: "🔬 Unit Test Login"
-        command: ".buildkite/commands/run-unit-tests.sh login"
+        command: ".buildkite/commands/run-unit-tests.sh" "login"
         plugins:
           - $CI_TOOLKIT
           - $TEST_COLLECTOR:
@@ -130,7 +130,7 @@ steps:
   # - group: "🔬 Instrumented tests"
   #   steps:
   #     - label: ":wordpress: 🔬 Instrumented tests"
-  #       command: ".buildkite/commands/run-instrumented-tests.sh wordpress"
+  #       command: ".buildkite/commands/run-instrumented-tests.sh" "wordpress"
   #       plugins:
   #         - $CI_TOOLKIT
   #         - $TEST_COLLECTOR :
@@ -140,7 +140,7 @@ steps:
   #         - "**/build/instrumented-tests/**/*"
 
   #     - label: ":jetpack: 🔬 Instrumented tests"
-  #       command: ".buildkite/commands/run-instrumented-tests.sh jetpack"
+  #       command: ".buildkite/commands/run-instrumented-tests.sh" "jetpack"
   #       plugins:
   #         - $CI_TOOLKIT
   #         - $TEST_COLLECTOR :
@@ -155,11 +155,11 @@ steps:
   - group: "📲 Prototype Builds"
     steps:
       - label: ":wordpress: :android: Prototype Build"
-        command: ".buildkite/commands/prototype-build.sh wordpress"
+        command: ".buildkite/commands/prototype-build.sh" "wordpress"
         if: build.pull_request.id != null
         plugins: [$CI_TOOLKIT]
 
       - label: ":jetpack: :android: Prototype Build"
-        command: ".buildkite/commands/prototype-build.sh jetpack"
+        command: ".buildkite/commands/prototype-build.sh" "jetpack"
         if: build.pull_request.id != null
         plugins: [$CI_TOOLKIT]